--- bsx3/ui/app/Main.jsx.orig
@@ -1,65 +1,0 @@
-import React, { Component } from 'react';
-<<<<<<< HEAD
-import { connect } from 'react-redux';
-import { Route } from 'react-router-dom';
-import { createBrowserHistory } from 'history';
-import NavLoggedOut from '../views/header/NavLoggedOut';
-import NavLoggedIn from '../views/header/NavLoggedIn';
-import Login from '../views/login/Login';
-import Counter from '../counter/Counter';
-import Datacollection from '../views/datacollection/DataCollection';
-import Notification from '../components/notification/Notification';
-import './main.css';
-=======
-import Container from 'react-bootstrap/Container';
-import NavLoggedIn from '../header/NavLoggedIn';
->>>>>>> upstream/master
-
-// import DataCollection from '../datacollection/DataCollection';
-import BeamlineStatus from '../beamlinestatus/BeamlineStatus';
-
-
-class Main extends Component {
-  render() {
-    return (
-      <div className="main">
-<<<<<<< HEAD
-        {this.props.showNotification
-          ? (
-            <Notification
-              level="success"
-              message="Succesfully delete row"
-              visible={this.props.showNotification}
-              handleShowNotification={this.props.handleShowNotification}
-            />
-          )
-          : null
-        }
-        { <Route exact path="/login" component={NavLoggedOut} />
-        }
-        { <Route exact path="/datacollection" component={NavLoggedIn} />
-        }
-        <Route exact path="/login" component={Login} />
-        <Route exact path="/counter" component={Counter} />
-        <Route exact path="/datacollection" component={Datacollection} />
-        <Route exact path="/" component={NavLoggedIn} />
-        <Route exact path="/" component={Datacollection} />
-=======
-        <Container fluid style={{ padding: '0px' }}>
-          <NavLoggedIn />
-          <BeamlineStatus key="bmstatatus" />
-        </Container>
->>>>>>> upstream/master
-      </div>
-    );
-  }
-}
-
-function mapStateToProps() {
-  return {
-  };
-}
-
-export default connect(
-  mapStateToProps,
-)(Main);
