+++ bsx3/ui/views/main/Main.jsx
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Route } from 'react-router-dom';
 import { createBrowserHistory } from 'history';
-import NavLoggedOut from '../views/header/NavLoggedOut';
-import NavLoggedIn from '../views/header/NavLoggedIn';
-import Login from '../views/login/Login';
-import Datacollection from '../views/datacollection/DataCollection';
-import Notification from '../components/notification/Notification';
+import NavLoggedOut from '../header/NavLoggedOut';
+import NavLoggedIn from '../header/NavLoggedIn';
+import Login from '../login/Login';
+import Datacollection from '../datacollection/DataCollection';
+import Notification from '../../components/notification/Notification';
 // import './main.css';
 
 const history = createBrowserHistory();
